refactor(codecommit): extract 886-line helper layer out of service.rs
service.rs was 5277 lines. Move the two free-function helper regions (response
wrappers, git-object hashing/SHA-1/SHA-256, ARN + enum validators, commit/tree
builders, merge three-way resolution, JSON serializers, plus the Tree type
alias) into a codecommit_helpers.rs sibling module, leaving service.rs focused
on the request handlers and the AwsService impl.

Helpers are now pub(crate) and glob-imported; the inline #[cfg(test)] modules
stay put (one super::normalize_rule_content ref repointed at the new module).
Pure code movement, no behavior change: service.rs drops to 4404 lines, the
181-fn set is identical before/after, all 23 unit tests pass, workspace clippy
-D warnings clean.
